fixed yz output, now int16_t

+++ main/drivers/include/mag.h
@@ -9,10 +9,10 @@
 int16_t get_x_output(i2c_inst_t *i2c);
 uint16_t get_x_output_raw(i2c_inst_t *i2c);
 
-int get_y_output(i2c_inst_t *i2c);
+int16_t get_y_output(i2c_inst_t *i2c);
 uint16_t get_y_output_raw(i2c_inst_t *i2c);
 
-int get_z_output(i2c_inst_t *i2c);
+int16_t get_z_output(i2c_inst_t *i2c);
 uint16_t get_z_output_raw(i2c_inst_t *i2c);
 
 int get_temp_output(i2c_inst_t *i2c);
